requirements.txt
Django
python-dotenv
psycopg2
pillow
ipython
djangorestframework
django-phonenumber-field
phonenumbers
django_filter
djangorestframework-simplejwt
coverage
